move server files into cmd
