+++ cmd/server/README.MD
@@ -1,0 +1,5 @@
+# R2 Publication Server
+
+This will be a production-ready publication streaming server.
+
+More documentation coming soon!
