+++ cmd/server/configs/.gitignore
