+++ cmd/server/main.go
@@ -5,9 +5,8 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/readium/r2-streamer-go/internal/config"
-	"github.com/readium/r2-streamer-go/internal/logging"
-	"github.com/readium/r2-streamer-go/pkg/server"
+	"github.com/readium/r2-streamer-go/cmd/server/internal/config"
+	"github.com/readium/r2-streamer-go/cmd/server/internal/logging"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
@@ -20,7 +19,7 @@ func init() {
 
 func main() {
 	bind := fmt.Sprintf("%s:%d", viper.GetString("bind-address"), viper.GetInt("bind-port"))
-	conf := server.ServerConfig{
+	conf := ServerConfig{
 		Bind:            bind,
 		Origins:         viper.GetStringSlice("origins"),
 		PublicationPath: viper.GetString("publication-path"),
@@ -28,7 +27,7 @@ func main() {
 		SentryDSN:       viper.GetString("sentry-dsn"),
 		CacheDSN:        viper.GetString("cache-dsn"),
 	}
-	s := server.NewR2GoServer(conf)
+	s := NewR2GoServer(conf)
 
 	server := &http.Server{
 		ReadTimeout:    10 * time.Second,
