+++ cmd/server/opds.go
@@ -1,4 +1,4 @@
-package server
+package main
 
 import (
 	"github.com/jinzhu/copier"
