+++ cmd/server/structs.go
@@ -1,4 +1,4 @@
-package server
+package main
 
 import (
 	"time"
